EDU-395 refactor: OpenAI Client에 맞추어서 Key 호출부 및 응답 요청부 수정
# Conflicts:
#	app/main.py
