+++ app/clients/openai_client.py
@@ -1,5 +1,11 @@
-import openai
+import os
 
+import openai
+from dotenv import load_dotenv
+
+
+load_dotenv()
+openai.api_key = os.getenv("OPENAI_API_KEY")
 
 class OpenAiClient:
     def __init__(self, model="gpt-4o", temperature=0.7, max_tokens=800):
