+++ app/routers/recommend_router.py
@@ -1,16 +1,13 @@
 from fastapi import APIRouter, HTTPException
-from pydantic import BaseModel
 from langchain_community.vectorstores import Chroma
 from langchain_community.embeddings import OpenAIEmbeddings
 from typing import List, Dict
 import os
-from dotenv import load_dotenv
-import openai
 
+from app.clients import ai_client
 from app.models.recommend.request import RecommendRequest
 from app.models.recommend.response import RecommendResponse
 
-load_dotenv()
 
 router = APIRouter()
 
@@ -18,35 +15,23 @@ router = APIRouter()
 #  GPT를 통한 AI Pick 결정 함수
 def call_gpt_rerank(contents: List[Dict], query: str, context: str) -> int:
     prompt = f"""
-다음은 사용자의 학습 맥락입니다:
-{context}
-
-사용자의 학습 목표와 관련하여, 아래 콘텐츠들 중 가장 적합한 하나를 골라주세요:
-{contents}
-
-가장 적합한 콘텐츠의 번호(index)를 하나의 숫자로만 출력해주세요.
+        다음은 사용자의 학습 맥락입니다:
+        {context}
+        
+        사용자의 학습 목표와 관련하여, 아래 콘텐츠들 중 가장 적합한 하나를 골라주세요:
+        {contents}
+        
+        가장 적합한 콘텐츠의 번호(index)를 하나의 숫자로만 출력해주세요.
     """
+
+    system_prompt = "당신은 교육 추천 전문가입니다."
     try:
-        response = openai.ChatCompletion.create(
-            model="gpt-4o",
-            messages=[
-                {"role": "system", "content": "당신은 교육 추천 전문가입니다."},
-                {"role": "user", "content": prompt}
-            ],
-            temperature=0.2,
-            max_tokens=10
-        )
-        idx_str = response.choices[0].message.content.strip()
+        idx_str = ai_client.create_chat_response(system_prompt, prompt).strip()
         return int(idx_str)
     except:
         return 0
 
-@router.post(
-    "/recommend",
-    response_model=List[RecommendResponse],
-    summary="개인화 콘텐츠 추천 API",
-    description="사전/사후 평가 및 진단 기반으로 사용자의 맥락에 맞는 콘텐츠 4개와 AI픽 1개를 제공합니다."
-)
+@router.post("/recommend", response_model=List[RecommendResponse], summary="개인화 콘텐츠 추천 API", description="사전/사후 평가 및 진단 기반으로 사용자의 맥락에 맞는 콘텐츠 4개와 AI픽 1개를 제공합니다.")
 async def recommend_content(req: RecommendRequest):
     try:
         openai_key = os.getenv("OPENAI_API_KEY")
@@ -93,5 +78,3 @@ async def recommend_content(req: RecommendRequest):
 
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Error during recommendation: {str(e)}")
-
-
